refactor(Table): merge checked #631
